% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameStand.R
\name{nameStand}
\alias{nameStand}
\title{Standardize scientific names}
\usage{
nameStand(nvec)
}
\arguments{
\item{nvec}{a vector of scientific names of plants of the Italian flora}
}
\value{
a dataframe with three columns, myname, closest_match, distance
}
\description{
Standardize a list of scientific names against the Checklist of the Italian flora
}
\examples{
my_names <- c("Crocus etruscus Parl.","Santolina pinnata Viv.")
s <- superparse(my_names)
}
